Fix Google Fonts missing font hinting info
Google fonts returns files that have the font hinting instructions
removed when the user agent is set as OSX. This is almost certainly due
to the fact that OSX disabled sub-pixel antialiasing by default back
with the release of Mojave.

Switching out the user agent for an equivalent windows one will ensure
that the font hinting is included, dramatically improving the rendering
of most fonts on 1x resolution screens where font hinting is enabled.

This seems to result in a roughly 20% increase in the size of the
fetched fonts.